[hdMaya] Moving texture utilities around and running formatting.
